Factor out steps builder into seperate file
